+++ src/pages/diagnosises/DStep3.less
@@ -1,22 +1,18 @@
-.dis{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 2.5rem;
-  background-color: rgba(119, 86, 55, 1);
+.f3{
+  position: fixed;
+  width: 90%;
+  margin: 0 5%;
+  display: flex;/* 如果把盒子设置为弹性盒子,必须要加的 */
+  flex-direction: row;/* 设置一个大盒子内容垂直排列*/
+  align-items: center;/* 设置一个大盒子内容居中排列*/
+  bottom: 20rpx;
 }
-.ddd{
-  font-size: 1rem;
-  color: #fff;
+.finish{
+  color:#DC143C;
+  border: 1px solid #DC143C;
 }
-.sta{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 2.5rem;
-}
-.next{
-
+.reback{
+  position: absolute;
+  color:	#228B22;
+  border: 1px solid 	#228B22;
 }
